requirements.txt
# Restaurant Trends ETL - Python Dependencies

# Data manipulation and Excel reading
pandas>=2.0.0
openpyxl>=3.1.0

# PostgreSQL database driver
psycopg2-binary>=2.9.0

# Environment variables and configuration
python-dotenv>=1.0.0
pyyaml>=6.0
